chore(db): mv unix_timestamp_milli to utils.rs
